review: misc unsigned-type tightening, strip cast comments, fix div-by-zero
DbOnTheRocks: bufferSize cast was on the wrong operand — (ulong)0.35
truncates to 0 and divides by zero. Restore master's
(ulong)(targetFileSize / CompressibilityHint).

Era1 (EraExporter/EraImporter/EraReader): drop the "Cast to long/int is
safe" comments — the casts are still there, but the prose is noise.
totalProcessed → ulong (no more (ulong)Interlocked.Increment cast).

Era1 EraWriter: _startNumber long → ulong; cast moved to the single
WriteInt64 call where the disk format is signed.

Era1 tests (Era1ModuleTests / EraExporterTests): start/end → ulong end
to end, drop (ulong)start / (ulong)end casts at the call site.

EraE:
- Admin/RPC: ExportHistory/ImportHistory and admin_*EraHistory take
  ulong from/to; drops the negativity check and the (ulong)from cast.
- Archive/EraReader: e2.First/LastBlock are already ulong; drop the
  redundant (ulong) casts.
- Archive/EraWriter: write _startNumber via new WriteUInt64 helper
  (was WriteInt64 with (long) cast).
- E2Store/E2StoreReader: add ReadUInt64; load _startBlock directly,
  remove the signed round-trip + "negative starting block" check.
- Export/EraExporter: epoch indices and totalProcessed ulong; drop the
  (ulong)EraWriter.MaxEraSize cast (already ulong-compatible).
- Proofs/Validator: GetAccumulatorForEpoch takes ulong; drop the
  (long)(blockNumber / (ulong)SlotsPerHistoricalRoot) cast pair.
- Store/EraStore + RemoteEraStoreDecorator: strip the "Boundary cast —
  safe" / "uint promotes cleanly" comments and the `int → uint` migration
  preamble. Decorator's manifest keys / epoch locals: int → uint, so
  GetBlockRangeAsync stops doing `(ulong)minEpoch * _maxEraSize`.
- Store/HttpRemoteEraClient + tests: manifest keys uint; TryParseEpoch
  outputs uint (drops the `&& epoch >= 0` check that was structurally
  impossible to fail).

EthStats: SkipLocalsInit on the three methods that stackalloc and only
read the prefix they wrote (TryParse, TryParseUInt64String,
TryParseInt64String). Required AllowUnsafeBlocks in the csproj.

Ethash.Test/DifficultyCalculatorTests: blocksAbove long → ulong, drop
the (ulong)blocksAbove cast at the Calculate call sites; TestCase
literals end with UL.

Evm.Benchmark/BlockProcessingBenchmark: startNonce + nonce int → ulong;
WithNonce takes startNonce + (ulong)i instead of the cast-the-sum form.

Evm.Precompiles/Bls12381G{1,2}Msm + PairingCheck + Eip2537:
DiscountForG1/G2 now ulong-in / ulong-out; k is ulong; DataGasCost loses
the (ulong)k * (ulong)Eip2537.DiscountForGn(k) cast pair.

Evm.Test:
- Eip1108Tests: _blockNumberAdjustment int → long (Istanbul-relative
  signed offset); expectedPrecompileGas long → ulong; drops the
  (ulong)expectedPrecompileGas cast at AssertGas sites.
- Tracing/GasEstimationTests: errorMargin int → ulong (matches
  GasEstimator.Estimate's ulong errorMargin parameter); TestCase
  -1 reworked to ulong.MaxValue ("out of bounds" still triggers).
- Tracing/ParityLikeTxTracerTests: drop (long)tx.GasLimit and
  (UInt256)tx.Value — both already match the assertion target type.

+++ src/Nethermind/Nethermind.Era1.Test/Era1ModuleTests.cs
@@ -297,14 +297,14 @@ public class Era1ModuleTests
         }
     }
 
-    [TestCase(true, 0L, 0L, 1000UL, 1001L, 9999L)]
-    [TestCase(true, 0L, 2000L, 1000UL, 1001L, 2000L)]
-    [TestCase(true, 3000L, 0L, 5000UL, 5001L, 9999L)]
-    [TestCase(true, 0L, 0L, 0UL, null, 0L)]
-    [TestCase(false, 0L, 0L, 0UL, 1L, 9999L)]
-    [TestCase(false, 0L, 0L, 2000UL, 2001L, 9999L)]
+    [TestCase(true, 0UL, 0UL, 1000UL, 1001UL, 9999UL)]
+    [TestCase(true, 0UL, 2000UL, 1000UL, 1001UL, 2000UL)]
+    [TestCase(true, 3000UL, 0UL, 5000UL, 5001UL, 9999UL)]
+    [TestCase(true, 0UL, 0UL, 0UL, null, 0UL)]
+    [TestCase(false, 0UL, 0UL, 0UL, 1UL, 9999UL)]
+    [TestCase(false, 0UL, 0UL, 2000UL, 2001UL, 9999UL)]
     [CancelAfter(120_000)] // macOS ARM runners need more time for 10k-block chain
-    public async Task EraExportAndImport(bool fastSync, long start, long end, ulong headBlockNumber, long? expectedMinSuggestedBlock, long expectedMaxSuggestedBlock, CancellationToken cancellationToken)
+    public async Task EraExportAndImport(bool fastSync, ulong start, ulong end, ulong headBlockNumber, ulong? expectedMinSuggestedBlock, ulong expectedMaxSuggestedBlock, CancellationToken cancellationToken)
     {
         const int ChainLength = 10000;
         await using IContainer outCtx = await EraTestModule.CreateExportedEraEnvWithCompleteBlockBuilder(ChainLength, cancellationToken: cancellationToken);
@@ -331,8 +331,8 @@ public class Era1ModuleTests
             })
             .AddSingleton<IEraConfig>(new EraConfig()
             {
-                From = (ulong)start,
-                To = (ulong)end,
+                From = start,
+                To = end,
                 ImportDirectory = tmpDir,
                 TrustedAccumulatorFile = Path.Join(tmpDir, EraExporter.AccumulatorFileName),
                 MaxEra1Size = 16,
