+++ src/Nethermind/Nethermind.Ethash.Test/DifficultyCalculatorTests.cs
@@ -51,28 +51,28 @@ namespace Nethermind.Ethash.Test
         }
 
         // previous difficulty bomb +  InitialDifficultyBombBlock + offset
-        [TestCase(9000000 + EthashDifficultyCalculator.InitialDifficultyBombBlock + 1)]
-        [TestCase(9000000 + EthashDifficultyCalculator.InitialDifficultyBombBlock + 3)]
-        [TestCase(9000000 + EthashDifficultyCalculator.InitialDifficultyBombBlock + 730000)]
-        public void London_calculation_should_not_be_equal_to_Berlin(long blocksAbove) => Calculation_should_not_be_equal_on_different_difficulty_hard_forks(blocksAbove,
+        [TestCase(9000000UL + EthashDifficultyCalculator.InitialDifficultyBombBlock + 1)]
+        [TestCase(9000000UL + EthashDifficultyCalculator.InitialDifficultyBombBlock + 3)]
+        [TestCase(9000000UL + EthashDifficultyCalculator.InitialDifficultyBombBlock + 730000)]
+        public void London_calculation_should_not_be_equal_to_Berlin(ulong blocksAbove) => Calculation_should_not_be_equal_on_different_difficulty_hard_forks(blocksAbove,
                 Berlin.Instance, London.Instance);
 
         // previous difficulty bomb +  InitialDifficultyBombBlock + offset
-        [TestCase(9700000 + EthashDifficultyCalculator.InitialDifficultyBombBlock + 1)]
-        [TestCase(9700000 + EthashDifficultyCalculator.InitialDifficultyBombBlock + 3)]
-        [TestCase(9700000 + EthashDifficultyCalculator.InitialDifficultyBombBlock + 730000)]
-        public void ArrowGlacier_calculation_should_not_be_equal_to_London0(long blocksAbove) => Calculation_should_not_be_equal_on_different_difficulty_hard_forks(blocksAbove,
+        [TestCase(9700000UL + EthashDifficultyCalculator.InitialDifficultyBombBlock + 1)]
+        [TestCase(9700000UL + EthashDifficultyCalculator.InitialDifficultyBombBlock + 3)]
+        [TestCase(9700000UL + EthashDifficultyCalculator.InitialDifficultyBombBlock + 730000)]
+        public void ArrowGlacier_calculation_should_not_be_equal_to_London0(ulong blocksAbove) => Calculation_should_not_be_equal_on_different_difficulty_hard_forks(blocksAbove,
                 London.Instance, ArrowGlacier.Instance);
 
         // previous difficulty bomb +  InitialDifficultyBombBlock + offset
-        [TestCase(10700000 + EthashDifficultyCalculator.InitialDifficultyBombBlock + 1)]
-        [TestCase(10700000 + EthashDifficultyCalculator.InitialDifficultyBombBlock + 3)]
-        [TestCase(10700000 + EthashDifficultyCalculator.InitialDifficultyBombBlock + 730000)]
-        public void GrayGlacier_calculation_should_not_be_equal_to_ArrowGlacier(long blocksAbove) => Calculation_should_not_be_equal_on_different_difficulty_hard_forks(blocksAbove,
+        [TestCase(10700000UL + EthashDifficultyCalculator.InitialDifficultyBombBlock + 1)]
+        [TestCase(10700000UL + EthashDifficultyCalculator.InitialDifficultyBombBlock + 3)]
+        [TestCase(10700000UL + EthashDifficultyCalculator.InitialDifficultyBombBlock + 730000)]
+        public void GrayGlacier_calculation_should_not_be_equal_to_ArrowGlacier(ulong blocksAbove) => Calculation_should_not_be_equal_on_different_difficulty_hard_forks(blocksAbove,
                 ArrowGlacier.Instance, GrayGlacier.Instance);
 
         private void Calculation_should_not_be_equal_on_different_difficulty_hard_forks(
-            long blocksAbove, IReleaseSpec firstHardfork, IReleaseSpec secondHardfork)
+            ulong blocksAbove, IReleaseSpec firstHardfork, IReleaseSpec secondHardfork)
         {
             UInt256 parentDifficulty = 0x55f78f7;
             ulong parentTimestamp = 1613570258;
@@ -80,12 +80,12 @@ namespace Nethermind.Ethash.Test
             ISpecProvider firstHardForkSpecProvider = Substitute.For<ISpecProvider>();
             firstHardForkSpecProvider.GetSpec(Arg.Any<ForkActivation>()).Returns(firstHardfork);
             EthashDifficultyCalculator firstHardforkDifficultyCalculator = new(firstHardForkSpecProvider);
-            UInt256 firstHardforkResult = firstHardforkDifficultyCalculator.Calculate(parentDifficulty, parentTimestamp, currentTimestamp, (ulong)blocksAbove, false);
+            UInt256 firstHardforkResult = firstHardforkDifficultyCalculator.Calculate(parentDifficulty, parentTimestamp, currentTimestamp, blocksAbove, false);
 
             ISpecProvider secondHardforkSpecProvider = Substitute.For<ISpecProvider>();
             secondHardforkSpecProvider.GetSpec(Arg.Any<ForkActivation>()).Returns(secondHardfork);
             EthashDifficultyCalculator secondHardforkDifficultyCalculator = new(secondHardforkSpecProvider);
-            UInt256 secondHardforkResult = secondHardforkDifficultyCalculator.Calculate(parentDifficulty, parentTimestamp, currentTimestamp, (ulong)blocksAbove, false);
+            UInt256 secondHardforkResult = secondHardforkDifficultyCalculator.Calculate(parentDifficulty, parentTimestamp, currentTimestamp, blocksAbove, false);
 
             Assert.That(secondHardforkResult, Is.Not.EqualTo(firstHardforkResult));
         }
